+++ src/main/java/com/github/zerkath/rosemary/DataTypes/Pieces.java
@@ -1,20 +1,20 @@
 package com.github.zerkath.rosemary.DataTypes;
 
 public class Pieces {
-  public static final int EMPTY = 0b00000;
-  public static final int PAWN = 0b00001;
-  public static final int KNIGHT = 0b00010;
-  public static final int BISHOP = 0b00011;
-  public static final int ROOK = 0b00100;
-  public static final int QUEEN = 0b00101;
-  public static final int KING = 0b00110;
-  public static final int WHITE = 0b10000;
-  public static final int BLACK = 0b01000;
+  public static final byte EMPTY = 0b00000;
+  public static final byte PAWN = 0b00001;
+  public static final byte KNIGHT = 0b00010;
+  public static final byte BISHOP = 0b00011;
+  public static final byte ROOK = 0b00100;
+  public static final byte QUEEN = 0b00101;
+  public static final byte KING = 0b00110;
+  public static final byte WHITE = 0b10000;
+  public static final byte BLACK = 0b01000;
 
-  public static final int m_type = 0b00111;
-  public static final int m_white = 0b10000;
-  public static final int m_black = 0b01000;
-  public static final int m_colour = m_white | m_black;
+  public static final byte m_type = 0b00111;
+  public static final byte m_white = 0b10000;
+  public static final byte m_black = 0b01000;
+  public static final byte m_colour = m_white | m_black;
 
   public static char getPromotion(short header) {
     char result =
@@ -28,7 +28,7 @@ public class Pieces {
     return (header & 0b1000) != 0 ? Character.toUpperCase(result) : result;
   }
 
-  public static int getPromotionNum(short header) {
+  public static byte getPromotionNum(short header) {
     int result =
         switch (header & MoveUtil.atomicMask) {
           case 0b111 -> QUEEN;
@@ -37,7 +37,7 @@ public class Pieces {
           case 0b100 -> KNIGHT;
           default -> '\0';
         };
-    return (header & 0b1000) != 0 ? result | WHITE : result | BLACK;
+    return (byte) ((header & 0b1000) != 0 ? result | WHITE : result | BLACK);
   }
 
   public static char getChar(int piece) {
@@ -56,16 +56,16 @@ public class Pieces {
     return result;
   }
 
-  public static int getNum(char piece) {
-    int result = Character.isUpperCase(piece) ? WHITE : BLACK;
+  public static byte getNum(char piece) {
+    byte result = Character.isUpperCase(piece) ? WHITE : BLACK;
     piece = Character.toLowerCase(piece);
     return switch (piece) {
-      case 'p' -> result | PAWN;
-      case 'n' -> result | KNIGHT;
-      case 'b' -> result | BISHOP;
-      case 'r' -> result | ROOK;
-      case 'q' -> result | QUEEN;
-      case 'k' -> result | KING;
+      case 'p' -> (byte) (result | PAWN);
+      case 'n' -> (byte) (result | KNIGHT);
+      case 'b' -> (byte) (result | BISHOP);
+      case 'r' -> (byte) (result | ROOK);
+      case 'q' -> (byte) (result | QUEEN);
+      case 'k' -> (byte) (result | KING);
       default -> 0;
     };
   }
